Move makeList to utils directory
